packages/sima-prompts/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sima-prompts"
version = "0.1.0"
description = "Prompt registry and renderer for SIMA cognitive modules"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
keywords = ["prompts", "llm", "templates"]
authors = [
    { name = "SIMA Team" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["sima_prompts"]
